split tests into 2 files for , - simple structure - more nested structures
